@@ -6,19 +6,20 @@
   "author": "Steve Fisher",
   "license": "ISC",
   "dependencies": {
-    "axios": "^0.22.0",
-    "bootstrap": "^5.1.1",
-    "express": "^4.17.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-markdown": "^7.0.1",
-    "react-redux": "^7.2.5",
-    "react-router-dom": "^5.3.0",
-    "react-scripts": "4.0.3",
-    "reactstrap": "^8.10.0",
-    "redux": "^4.1.1",
+    "axios": "^1.2.5",
+    "bootstrap": "^5.2.3",
+    "express": "^4.18.2",
+    "npm-check-updates": "^16.6.3",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
+    "react-markdown": "^8.0.5",
+    "react-redux": "^8.0.5",
+    "react-router-dom": "^6.7.0",
+    "react-scripts": "5.0.1",
+    "reactstrap": "^9.1.5",
+    "redux": "^4.2.0",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0"
+    "redux-thunk": "^2.4.2"
   },
   "scripts": {
     "start": "react-scripts start",
