@@ -2,7 +2,7 @@ import React, { useEffect } from "react";
 import IssueTracker from "../../components/IssueTracker/IssueTracker";
 import { Row, Col, Card, CardText, CardSubtitle, Button } from "reactstrap";
 import { useDispatch, useSelector } from "react-redux";
-import { Link, useHistory } from "react-router-dom";
+import { Link, useNavigate } from "react-router-dom";
 import { listIssuesAction } from "../../redux/actions/issuesActions";
 import Loading from "../Loading";
 import ErrorMessage from "../ErrorMessage";
@@ -11,7 +11,7 @@ import "./CurrentIssue.css";
 
 function CurrentIssue({ search }) {
   const dispatch = useDispatch();
-  const history = useHistory();
+  const navigate = useNavigate();
 
   const currentIssues = useSelector((state) => state.currentIssues);
   const { loading, issues, error } = currentIssues;
@@ -28,9 +28,9 @@ function CurrentIssue({ search }) {
   useEffect(() => {
     dispatch(listIssuesAction());
     if (!userInfo) {
-      history.push("/");
+      navigate("/");
     }
-  }, [dispatch, successCreate, history, userInfo, successUpdate]);
+  }, [dispatch, navigate, successCreate, userInfo, successUpdate]);
 
   return (
     <div className="issue-container">
@@ -50,7 +50,7 @@ function CurrentIssue({ search }) {
                   .toLowerCase()
                   .includes(search.toLowerCase())
               )
-              .map((issue) => (
+              .reverse().map((issue) => (
                 <Card key={issue._id} className="mb-2">
                   <Row className="pt-1 px-2">
                     <Col>
@@ -77,7 +77,7 @@ function CurrentIssue({ search }) {
                   </Row>
                   <Row className="pt-1 px-2">
                     <CardSubtitle className="text-left">
-                      Description
+                      Description:
                     </CardSubtitle>
                   </Row>
                   <Row className="pt-2">
@@ -87,8 +87,7 @@ function CurrentIssue({ search }) {
                   </Row>
                   <Row className="pt-2">
                     <footer className="footer">
-                      Created on {issue.createdAt.substring(0, 10)} / Updated on
-                      {issue.updatedAt.substring(0, 10)}
+                      Created on: {issue.createdAt.substring(0, 10)} / Updated on: {issue.updatedAt.substring(0, 10)}
                     </footer>
                   </Row>
                 </Card>
