@@ -1,20 +1,20 @@
 import React, { useState, useEffect } from "react";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 import { Alert } from "reactstrap";
 
 const ErrorMessage = ({ color = "info", children }) => {
   const [visible, setVisible] = useState(true);
-  const history = useHistory();
+  const navigate = useNavigate();
 
   useEffect(() => {
     const timeId = setTimeout(() => {
       setVisible(false);
-      history.push("/issues");
+      navigate("/issues");
     }, 5000);
     return () => {
       clearTimeout(timeId);
     };
-  }, [history]);
+  }, [navigate]);
 
   if (!visible) {
     return null;
