@@ -11,7 +11,7 @@ import {
   USER_UPDATE_SUCCESS,
 } from "../constants/userConstants";
 import axios from "axios";
-import { useHistory } from "react-router-dom";
+import { useNavigate } from "react-router-dom";
 
 export const login = (email, password) => async (dispatch) => {
   try {
@@ -44,8 +44,10 @@ export const login = (email, password) => async (dispatch) => {
 };
 
 export const logout = () => async (dispatch) => {
+  const navigate = useNavigate();
   localStorage.removeItem("userInfo");
   dispatch({ type: USER_LOGOUT });
+  navigate("/")
 };
 
 export const register =
@@ -72,9 +74,10 @@ export const register =
       localStorage.setItem("userInfo", JSON.stringify(data));
 
       const userInfo = localStorage.getItem("userInfo");
-      const history = useHistory();
+      const navigate = useNavigate();
+
       if (userInfo) {
-        history.push("/issues");
+        navigate("/issues");
       }
     } catch (error) {
       dispatch({
